Do minor refactoring of svg context
