+++ draw2dsvg/gc.go
@@ -5,13 +5,9 @@ package draw2dsvg
 
 import (
 	"bytes"
-	"fmt"
 	"github.com/llgcode/draw2d"
 	"github.com/llgcode/draw2d/draw2dbase"
 	"image"
-	"image/color"
-	"math"
-	"strings"
 )
 
 const ()
@@ -28,7 +24,10 @@ const (
 type SVG bytes.Buffer
 
 func NewSvg() *Svg {
-	return &Svg{Xmlns: "http://www.w3.org/2000/svg"}
+	return &Svg{
+		Xmlns:      "http://www.w3.org/2000/svg",
+		FillStroke: FillStroke{Fill: "none", Stroke: "none"},
+	}
 }
 
 // GraphicContext implements the draw2d.GraphicContext interface
@@ -74,109 +73,30 @@ func (gc *GraphicContext) drawPaths(drawType drawType, paths ...*draw2d.Path) {
 
 	svgPaths := make([]Path, len(paths))
 
+	group := Group{
+		Paths: svgPaths,
+	}
+
 	for i, path := range paths {
 		svgPaths[i].Desc = toSvgPathDesc(path)
-		if drawType&stroked == stroked {
-			svgPaths[i].Stroke = toSvgRGBA(gc.Current.StrokeColor)
-			svgPaths[i].StrokeWidth = toSvgLength(gc.Current.LineWidth)
-			svgPaths[i].StrokeLinecap = gc.Current.Cap.String()
-			svgPaths[i].StrokeLinejoin = gc.Current.Join.String()
-			if len(gc.Current.Dash) > 0 {
-				svgPaths[i].StrokeDasharray = toSvgArray(gc.Current.Dash)
-				svgPaths[i].StrokeDashoffset = toSvgLength(gc.Current.DashOffset)
-			}
-		} else {
-			svgPaths[i].Stroke = "none"
-		}
-		if drawType&filled == filled {
-			svgPaths[i].Fill = toSvgRGBA(gc.Current.FillColor)
-		} else {
-			svgPaths[i].Fill = "none"
+	}
+
+	if drawType&stroked == stroked {
+		group.Stroke = toSvgRGBA(gc.Current.StrokeColor)
+		group.StrokeWidth = toSvgLength(gc.Current.LineWidth)
+		group.StrokeLinecap = gc.Current.Cap.String()
+		group.StrokeLinejoin = gc.Current.Join.String()
+		if len(gc.Current.Dash) > 0 {
+			group.StrokeDasharray = toSvgArray(gc.Current.Dash)
+			group.StrokeDashoffset = toSvgLength(gc.Current.DashOffset)
 		}
 	}
 
-	gc.svg.Groups = append(gc.svg.Groups, Group{
-		Paths: svgPaths,
-	})
-}
-
-func toSvgRGBA(c color.Color) string { // TODO move elsewhere
-	r, g, b, a := c.RGBA()
-	return fmt.Sprintf("rgba(%v, %v, %v, %.3f)", r>>8, g>>8, b>>8, float64(a>>8)/255)
-}
-
-func toSvgLength(l float64) string {
-	return fmt.Sprintf("%.4f", l)
-}
-
-func toSvgArray(nums []float64) string {
-	arr := make([]string, len(nums))
-	for i, num := range nums {
-		arr[i] = fmt.Sprintf("%.4f", num)
+	if drawType&filled == filled {
+		group.Fill = toSvgRGBA(gc.Current.FillColor)
 	}
-	return strings.Join(arr, ",")
-}
 
-func toSvgPathDesc(p *draw2d.Path) string { // TODO move elsewhere
-	parts := make([]string, len(p.Components))
-	ps := p.Points
-	for i, cmp := range p.Components {
-		switch cmp {
-		case draw2d.MoveToCmp:
-			parts[i] = fmt.Sprintf("M %.4f,%.4f", ps[0], ps[1])
-			ps = ps[2:]
-		case draw2d.LineToCmp:
-			parts[i] = fmt.Sprintf("L %.4f,%.4f", ps[0], ps[1])
-			ps = ps[2:]
-		case draw2d.QuadCurveToCmp:
-			parts[i] = fmt.Sprintf("Q %.4f,%.4f %.4f,%.4f", ps[0], ps[1], ps[2], ps[3])
-			ps = ps[4:]
-		case draw2d.CubicCurveToCmp:
-			parts[i] = fmt.Sprintf("C %.4f,%.4f %.4f,%.4f %.4f,%.4f", ps[0], ps[1], ps[2], ps[3], ps[4], ps[5])
-			ps = ps[6:]
-		case draw2d.ArcToCmp:
-			cx, cy := ps[0], ps[1] // center
-			rx, ry := ps[2], ps[3] // radii
-			fi := ps[4] + ps[5]    // startAngle + angle
-
-			// compute endpoint
-			sinfi, cosfi := math.Sincos(fi)
-			nom := math.Hypot(ry*cosfi, rx*sinfi)
-			x := cx + (rx*ry*cosfi)/nom
-			y := cy + (rx*ry*sinfi)/nom
-
-			// compute large and sweep flags
-			large := 0
-			sweep := 0
-			if math.Abs(ps[5]) > math.Pi {
-				large = 1
-			}
-			if !math.Signbit(ps[5]) {
-				sweep = 1
-			}
-			// dirty hack to ensure whole arc is drawn
-			// if start point equals end point
-			if sweep == 1 {
-				x += 0.001 * sinfi
-				y += 0.001 * -cosfi
-			} else {
-				x += 0.001 * sinfi
-				y += 0.001 * cosfi
-			}
-
-			// rx ry x-axis-rotation large-arc-flag sweep-flag x y
-			parts[i] = fmt.Sprintf("A %.4f %.4f %v %v %v %.4f %.4f",
-				rx, ry,
-				0,
-				large, sweep,
-				x, y,
-			)
-			ps = ps[6:]
-		case draw2d.CloseCmp:
-			parts[i] = "Z"
-		}
-	}
-	return strings.Join(parts, " ")
+	gc.svg.Groups = append(gc.svg.Groups, group)
 }
 
 ///////////////////////////////////////
